hello 2023

+++ B_Equivalent_Strings.cpp
@@ -316,7 +316,7 @@ int main()
     //#endif
     ll t = 1;
     // cin >> t;
-    for (int it = 1; it <= t; it++)
+    for (ll it = 1; it <= t; it++)
     {
         solve();
     }
